Allow files as environment variables
Allows the use of container files as a replacement for environment
variables.
The same environment variable names are used, but with a '_FILE' suffix.
When defined and a valid / existing file, they take precedence over the
'normal' environment variable. When not defined, or not a valid file,
the 'normal' environment variable will be used, or when not defined the
default value (if one exists).

This allows for docker secrets that are mounted as files under
/run/secrets/{secret_name}.