create sqlite database file if it does not exist
When a volume mount is used for /var/wwww/wallabag/data, as
documented[^1], the container fails to start with the error messages:

    wc: /var/www/wallabag/data/db/wallabag.sqlite: No such file or directory
    Configuring the SQLite database ...
    In ExceptionConverter.php line 76:
      An exception occurred in the driver: SQLSTATE[HY000] [14] unable to
      open database file
    [..]

In the volume mount the db directory and the empty db/wallabag.sqlite do
not exist which causes the issue.

Check in the entrypoint script if the file exist, if it does not create
the db directory and the file and make nobody the owner.

This fixes: https://github.com/wallabag/docker/issues/316

[^1]:https://github.com/wallabag/docker/tree/96dd9bc3e3d12b216471db03f5afacd6d1afac61#sqlite